reactor : TournamentBracket 컴포넌트 depth 줄이기위해 결합했습니다
