--- src/1_Page/Championship/ui/DashBoard/ui/TournamentBracket/ui/TournamentMatchColumn/type.d.ts
@@ -1,3 +1,0 @@
-type TournamentMatchColumnProps = {
-  match: ChampionshipMatchList;
-};
