+++ src/2_Widget/PlayerHistoryTable/ui/PlayerHistoryRow/model/type.d.ts
