+++ src/2_Widget/PlayerHistoryTable/ui/PlayerHistoryRow/type.d.ts
